+++ bin/core/configuration.js
@@ -1,30 +1,50 @@
 const process = require('process')
-const nconf = require('nconf')
+const fs = require('fs-extra')
 const _ = require('lodash')
 const os = require('os')
 
-nconf.argv()
+const frozenConfiguration = {}
 
-_.each([
-  `configuration/standard.json`,
-  `${os.homedir()}/.bauble`,
-  `${process.cwd()}/.bauble`
-], (configurationPath, index) => {
-  try {
-    nconf.file(index, configurationPath)
-  } catch (e) {
-    ;
+class Configuration {
+  constructor() {
+    this.__configuration = {}
   }
-})
 
-// hi-jack the get method to implement depth searching on keys
-const _nconfGet = nconf.get
-nconf.get = function(path, defaultValue = ``) {
-  return (!!path)
-    ? _.get(_nconfGet.call(this), path, defaultValue)
-    : _nconfGet.call(this)
+  override(manifestJson) {
+    this.__configuration = _.merge({}, this.__configuration, manifestJson)
+  }
+
+  load() {
+
+    _.each([
+      `configuration/standard.json`,
+      `${os.homedir()}/.bauble`,
+      `${process.cwd()}/.bauble`
+    ], (configurationPath) => {
+      this.loadFile(configurationPath)
+    })
+
+    return this
+  }
+
+  loadFile(filename) {
+    try {
+      this.__configuration = _.merge(this.__configuration,
+        JSON.parse(fs.readFileSync(filename).toString() || '{}')
+      )
+    }
+    catch(e) {
+      ;
+    }
+  }
+
+  get(path, defaultValue = '') {
+    return (path)
+      ? _.get(this.__configuration, path, defaultValue)
+      : this.__configuration
+  }
 }
 
 module.exports = {
-  configuration: nconf
+  configuration: new Configuration().load()
 }